refactor: fix import paths and update folder structure
- Fix database schema imports from lib/db to lib/db/core/schema
- Fix realtime connection store to stores/connection
- Fix realtime schemas to schema
- Move realtime services to services/ subdirectory
- Fix stores imports (store -> stores)
- Fix type declarations (.d.ts files)
- Update documentation with new folder structure
