+++ src/lib/realtime/services/chat.ts
@@ -5,7 +5,7 @@
  */
 
 import { randomUUID } from "uncrypto";
-import { connectionStore } from "./connection-store";
+import { connectionStore } from "../stores/connection";
 import { logger } from "~/lib/logger";
 
 /**
