+++ test/lib/mcp/rate-limit.test.ts
@@ -10,7 +10,7 @@ import {
   cleanupRateLimitStore,
   memoryRateLimitStore,
 } from "../../../src/lib/mcp/rate-limit";
-import type { McpApiKey } from "../../../src/lib/db/schema";
+import type { McpApiKey } from "../../../src/lib/db/core/schema";
 
 /**
  * Mock API key factory.
